requirements.txt
# GUI
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0

# Base
python-dotenv==1.0.0
SQLAlchemy==2.0.25
pytest==7.4.4
setuptools==69.1.0

# AI y Speech
google.generativeai==0.3.1
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
selenium==4.18.1
webdriver-manager==4.0.1

# Media y Audio
pygame==2.5.2
yt-dlp==2023.12.30
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.14

# Utilidades
numpy==1.26.3
psutil==5.9.8
